- Metal implementation of the network test phase. Each GPU thread executes the inference on a different input data.
- Change non-signed integers from size_t to unsigned int since we often don't need a 64-bit integer.

+++ include/LoadMNISTDataSet.h
@@ -12,5 +12,5 @@
 
 #endif /* LoadMNISTDataSet_h */
 
-float * _Nonnull * _Nonnull loadMnist(const char * _Nonnull file, size_t * _Nonnull len1, size_t * _Nonnull len2);
-float * _Nonnull * _Nonnull loadMnistTest(const char * _Nonnull file, size_t * _Nonnull len1, size_t * _Nonnull len2);
+float * _Nonnull * _Nonnull loadMnist(const char * _Nonnull file, unsigned int * _Nonnull len1, unsigned int * _Nonnull len2);
+float * _Nonnull * _Nonnull loadMnistTest(const char * _Nonnull file, unsigned int * _Nonnull len1, unsigned int * _Nonnull len2);
